% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_token.R
\name{generate_token}
\alias{generate_token}
\title{Generate an access token}
\usage{
generate_token(client_id, client_secret, expiration = "60")
}
\arguments{
\item{client_id}{ID for the ArcGIS REST API}

\item{client_secret}{Secret for the ArcGIS REST API}

\item{expiration}{The length of time the token should last in minutes}
}
\description{
Generate an access token
}
\references{
https://developers.arcgis.com/
}
