% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drive_time.R
\name{get_drive_time}
\alias{get_drive_time}
\title{Get drive time between two points}
\usage{
get_drive_time(token, x1, y1, x2, y2)
}
\arguments{
\item{x1, y1, x2, y2}{Eastings and northings of two points}
}
\description{
Get drive time between two points
}
