% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drive_time_demographics.R
\name{get_drive_time_demographics}
\alias{get_drive_time_demographics}
\title{Get demographic data from ArcGIS}
\usage{
get_drive_time_demographics(
  token,
  x,
  y,
  drive_time = 10,
  unit = "Minutes",
  demographics = c("KeyGlobalFacts", "KeyUSFacts")
)
}
\arguments{
\item{token}{API access token for ArcGIS}

\item{x}{Easting (longitude)}

\item{y}{Northing (lattitude)}

\item{drive_time}{Distance in driving time}

\item{unit}{Drive time unit (Hours or Minutes)}

\item{demographics}{What demographic information (see Description)}
}
\description{
This function queries demographic and lifestyle information within
a drive time polygon from the ArcGIS World GeoEnrichmentService. See the
documentation here:
https://developers.arcgis.com/rest/geoenrichment/api-reference/enrich.htm
}
\examples{
\dontrun{
client_id <- "your client id"
client_secret <- "your client secret"

token <- generate_token(client_id, client_secret)

get_drive_time_demographics(token, -86.157963, 39.768454)
}
}
