% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_long.R
\name{get_lat_long}
\alias{get_lat_long}
\title{Get lat/long for address}
\usage{
get_lat_long(token, address, state, zip_code)
}
\arguments{
\item{address, state, zip_code}{Address for geocoding}
}
\description{
Get lat/long for address
}
